Move qsirecon workflow.
